fix: разнесены api сервисы по модулям
