+++ webapi/src/app/internal/api_services/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseAPIService
+from .tests import test_completed_service, test_progress_service
+from .users import user_service, user_time_slot_service
